+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/operations/listsyncs.ts
@@ -37,7 +37,7 @@ export class ListSyncsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success