+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/operations/updateexpensedataset.ts
@@ -59,7 +59,7 @@ export class UpdateExpenseDatasetResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Accepted