+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/operations/uploadattachment.ts
@@ -68,5 +68,5 @@ export class UploadAttachmentResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }