+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/webhooks/syncfailed.ts
@@ -22,5 +22,5 @@ export class SyncFailedResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }