+++ previous-versions/sync-for-expenses-version-1/src/sdk/sdk.ts
@@ -56,10 +56,10 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "prealpha";
-    sdkVersion = "0.5.0";
-    genVersion = "2.188.1";
+    sdkVersion = "0.5.1";
+    genVersion = "2.194.1";
     userAgent =
-        "speakeasy-sdk/typescript 0.5.0 2.188.1 prealpha @codat/sync-for-expenses-version-1";
+        "speakeasy-sdk/typescript 0.5.1 2.194.1 prealpha @codat/sync-for-expenses-version-1";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);